+++ backend/src/main/java/cn/lz/seq/service/SeqServiceImpl.java
@@ -4,6 +4,9 @@ import cn.lz.seq.api.SeqService;
 import cn.lz.seq.dao.SeqDao;
 import cn.lz.seq.zk.ZNodeWatcherService;
 import lombok.extern.slf4j.Slf4j;
+
+import java.math.BigInteger;
+
 import org.apache.dubbo.config.annotation.DubboService;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -17,13 +20,13 @@ public class SeqServiceImpl implements SeqService {
     @Autowired
     private ZNodeWatcherService zNodeWatcherService;
 
-    public long getSeq(String token) {
+    public BigInteger getSeq(String token) {
         String tableName = zNodeWatcherService.getTableNameByToken(token);
         if (tableName == null) {
             log.info("token: {}, seq: {}", token, -1);
-            return -1;
+            return BigInteger.valueOf(-1);
         }
-        long seq = seqDao.getSeq(tableName);
+        BigInteger seq = seqDao.getSeq(tableName);
         log.info("token: {}, seq: {}", token, seq);
         return seq;
     }
