+++ interface/src/main/java/cn/lz/seq/api/SeqService.java
@@ -1,5 +1,7 @@
 package cn.lz.seq.api;
 
+import java.math.BigInteger;
+
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,5 +11,5 @@ import org.springframework.web.bind.annotation.RestController;
 public interface SeqService {
 
     @RequestMapping("/seq")
-    long getSeq(@RequestParam("token") String token);
+    BigInteger getSeq(@RequestParam("token") String token);
 }
